+++ src/components/FixCreationTime/options.tsx
@@ -1,15 +1,9 @@
-import { Formik } from 'formik';
 import React, { ChangeEvent } from 'react';
 import { FIX_OPTIONS } from '.';
 import { Form } from 'react-bootstrap';
 import EnteDateTimePicker from 'components/EnteDateTimePicker';
 import { Row, Value } from 'components/Container';
 import constants from 'utils/strings/constants';
-
-interface formValues {
-    option: FIX_OPTIONS;
-    customTime: Date;
-}
 
 const Option = ({
     value,
@@ -43,65 +37,48 @@ const Option = ({
     </Form.Check>
 );
 
-export function FixCreationTimeOptions({ handleSubmit }) {
-    const onSubmit = (values: formValues) => {
-        console.log(values);
-        handleSubmit(values.option, new Date(values.customTime));
-    };
-
+export default function FixCreationTimeOptions({ handleChange, values }) {
     return (
-        <Formik<formValues>
-            initialValues={{
-                option: FIX_OPTIONS.DATE_TIME_ORIGINAL,
-                customTime: new Date(),
-            }}
-            validateOnBlur={false}
-            onSubmit={onSubmit}>
-            {({ values, handleChange, handleSubmit }) => (
-                <Form noValidate onSubmit={handleSubmit}>
-                    <Value width="40%">
-                        <Option
-                            value={FIX_OPTIONS.DATE_TIME_ORIGINAL}
-                            onChange={handleChange('option')}
-                            label={constants.CREATION_DATE_TIME}
-                            selected={Number(values.option)}
+        <Form noValidate>
+            <Value width="40%">
+                <Option
+                    value={FIX_OPTIONS.DATE_TIME_ORIGINAL}
+                    onChange={handleChange('option')}
+                    label={constants.CREATION_DATE_TIME}
+                    selected={Number(values.option)}
+                />
+            </Value>
+
+            <Value width="40%">
+                <Option
+                    value={FIX_OPTIONS.DATE_TIME_DIGITIZED}
+                    onChange={handleChange('option')}
+                    label={constants.DATE_TIME_DIGITIZED}
+                    selected={Number(values.option)}
+                />
+            </Value>
+
+            <Row style={{ margin: '0' }}>
+                <Value width="40%">
+                    <Option
+                        value={FIX_OPTIONS.CUSTOM_TIME}
+                        onChange={handleChange('option')}
+                        label={constants.CUSTOM_TIME}
+                        selected={Number(values.option)}
+                    />
+                </Value>
+                {Number(values.option) === FIX_OPTIONS.CUSTOM_TIME && (
+                    <Value width="40%%">
+                        <EnteDateTimePicker
+                            isInEditMode
+                            pickedTime={new Date(values.customTime)}
+                            handleChange={(x: Date) =>
+                                handleChange('customTime')(x.toUTCString())
+                            }
                         />
                     </Value>
-
-                    <Value width="40%">
-                        <Option
-                            value={FIX_OPTIONS.DATE_TIME_DIGITIZED}
-                            onChange={handleChange('option')}
-                            label={constants.DATE_TIME_DIGITIZED}
-                            selected={Number(values.option)}
-                        />
-                    </Value>
-
-                    <Row style={{ margin: '0' }}>
-                        <Value width="40%">
-                            <Option
-                                value={FIX_OPTIONS.CUSTOM_TIME}
-                                onChange={handleChange('option')}
-                                label={constants.CUSTOM_TIME}
-                                selected={Number(values.option)}
-                            />
-                        </Value>
-                        {Number(values.option) === FIX_OPTIONS.CUSTOM_TIME && (
-                            <Value width="40%%">
-                                <EnteDateTimePicker
-                                    isInEditMode
-                                    pickedTime={new Date(values.customTime)}
-                                    handleChange={(x: Date) =>
-                                        handleChange('customTime')(
-                                            x.toUTCString()
-                                        )
-                                    }
-                                />
-                            </Value>
-                        )}
-                    </Row>
-                </Form>
-            )}
-        </Formik>
+                )}
+            </Row>
+        </Form>
     );
 }
